[FIX] update the tax address properly
Since the tax_id field is readonly in certain cases you cannot rely on
the onchanges to handle this. This solves the issue by turning tax_id
into a stored computed field. Note that the various onchanges are
still necessary to manage the booleans (which should have been a
single Selection field).