[FIX] support suds-jurko 0.7
Although never published on pypi.org Debian Stretch' python3-suds
contains suds-jurko 0.7. It's not backwards-compatible with code
written for suds-jurko 0.6.

Contrary to 0.6, 0.7 respects minOccurs and maxOccurs attributes
which caused empty TaxOccurs elements to be inserted. This in turn
made Avatax return an error.

Additionally, 0.7 no longer maps a WSDL 'decimal' to a Python
float. Instead it maps it to a Python Decimal.

Fixes #6