[FIX] account_avatax: don't autovalidate non supported country addresses
With automatic address validation enabled, when an address for a
non-supported country was used, an error was displayed.

With this fix, addresses for non-supported countries will not be
subject to automatic validation.

For example: with automatic address validation enabled,
set Customer to an Europe address, and set the delievry
address to an US address. Computing taxes should apply US Sales Tax
and raise no error for the European address.

(cherry picked from commit 72c77dc35c28e8316456e7fef8fcdede945f96b2)